% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.init.r
\name{gam.init.sub}
\alias{gam.init.sub}
\title{Function to select initial gamma values spanning its space}
\usage{
gam.init.sub(q, n.initials)
}
\arguments{
\item{n.initials}{Number of gamma values.}

\item{Z}{The grouping variables.}

\item{lb.quantile}{The lower quantile specified for Z\%*\%gamma.initials.}

\item{ub.quantile}{The upper quantile specified for Z\%*\%gamma.initials.}

\item{ss}{A positive integer with n.initials/ss indicating how many sets of gamma are chosen, default to 1.}
}
\value{
A matrix. A set of gamma values selected for defining subgroup.
}
\description{
Function to select initial gamma values spanning its space
}
\examples{
n = 100
p1 = 2
p2 = 1
p3 = 3
alpha = rep(1, p1)
beta  = rep(1, p2)/2
gamma = c(1, seq(-1,1,length.out = p3-1)) 
rho = 0.3
cenRate = 0.2
data = generate_cox_data(n, alpha, beta, gamma, rho, cenRate = cenRate)
K = 1000
qlb = 0.1
cols = apply(data$u, 2, var) != 0
Gamma = gam.init(K, data$u[,cols], lb.quantile=qlb, ub.quantile=1-qlb, ss=1)
}
