% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cox_data.r
\name{generate_cox_data}
\alias{generate_cox_data}
\title{Function for generating data from Cox proportional hazard model with a change plane.}
\usage{
generate_cox_data(
  n,
  alpha,
  beta,
  gamma,
  rho,
  cenRate = 0.1,
  censortype = c("RightCensor", "RandomCensor")
)
}
\arguments{
\item{n}{A constant. The sample size.}

\item{alpha}{A vector. The true parameter for baseline covariates.}

\item{beta}{A vector. The true parameter denoting the heterogeneous effect of the subgroup.}

\item{gamma}{A vector. The true parameter for grouping variables.}

\item{rho}{A constant. The strength of correlation among covariates.}

\item{cenRate}{A constant. Censoring rate. Default is 0.1.}

\item{censortype}{Censroing type, including "RightCensor" and "RandomCensor".}
}
\value{
A list
\itemize{
\item y - A length \eqn{n} vector. The survival time.
\item x - A matrix. The baseline covariates.
\item z - A matrix. The baseline covariates.
\item u - A matrix. The grouping variables.
\item status - A length \eqn{n} vector. Censoring indicator.
}
}
\description{
Function for generating data from Cox proportional hazard model with a change plane.
}
\examples{
n = 100
p1 = 2
p2 = 1
p3 = 3
alpha = rep(1, p1)
beta  = rep(1, p2)/2
gamma = c(1, seq(-1,1,length.out = p3-1)) 
rho = 0.3
cenRate = 0.2
data = generate_cox_data(n, alpha, beta, gamma, rho, cenRate = cenRate)
}
